--- src/components/pages/Mypage/_common/index.js
@@ -1,3 +1,0 @@
-import MypageButton from "./MypageButton";
-
-export {MypageButton}
