+++ src/components/pages/Mypage/components/constants.js
@@ -1,0 +1,7 @@
+const constants = {
+  ARTICLE: 1,
+  COMMENT: 2,
+  LIKED: 3,
+};
+
+export default constants;
