+++ src/components/pages/Mypage/components/index.js
@@ -1,0 +1,5 @@
+import MyProfile from '../MyProfile';
+import MyArticlePreview from './MyArticlePreview';
+import MyArticleBoard from './MyArticleBoard';
+
+export { MyProfile, MyArticleBoard, MyArticlePreview };
