+++ src/components/pages/Mypage/styles/index.js
@@ -1,0 +1,3 @@
+import StyledMyPageButton from './Mypagebutton.styles';
+
+export { StyledMyPageButton };
